+++ lib/promise_pool/future.rb
@@ -1,6 +1,21 @@
 
 module PromisePool
   class Future < BasicObject
+    def self.resolve future
+      if future.kind_of?(::Array)
+        future.map(&method(:resolve))
+
+      elsif future.kind_of?(::Hash)
+        future.inject({}) do |r, (k, v)|
+          r[k] = resolve(v)
+          r
+        end
+
+      else
+        future.itself
+      end
+    end
+
     def initialize promise
       @promise = promise
     end