+++ CHANGES.md
@@ -1,5 +1,9 @@
 # CHANGES
 
+## promise_pool 0.9.1 -- 2016-07-21
+
+* Introduced PromisePool::Future.resolve to convert nested futures
+
 ## promise_pool 0.9.0 -- 2016-01-29
 
 * First beta!